date, string 유틸함수 통합
