+++ src/modules/postCard/index.tsx
@@ -8,7 +8,7 @@ import { PostIntroType } from '@/types/components';
 import Badge from '@/ui/badge';
 import Heading from '@/ui/heading';
 import Text from '@/ui/text';
-import { formatDate, toISODate } from '@/utils/core/format';
+import { formatDate, toISODate } from '@/utils/core/date';
 
 import * as styles from './styles.css';
