+++ src/utils/core/date.ts
@@ -13,9 +13,3 @@ export const formatDate = (date: Date, lang: Locale) => {
 
   return formatter.format(date);
 };
-
-export const slugify = (input: string) =>
-  input
-    .toLowerCase()
-    .replace(/\s+/g, '-')
-    .replace(/[^a-z0-9-]/g, '');
